pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "integral-superposition"
version = "0.1.0"
description = "Model-agnostic Sparse Autoencoder training, analysis, and steering toolkit"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Research Team", email = "research@example.com"},
]
keywords = ["sparse-autoencoders", "mechanistic-interpretability", "ml-interpretability"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=1.13.0",
    "transformers>=4.20.0",
    "numpy>=1.21.0",
    "pandas>=1.5.0",
    "scikit-learn>=1.1.0",
    "scipy>=1.9.0",
    "matplotlib>=3.5.0",
    "tqdm>=4.64.0",
    "pydantic>=1.10.0",
    "networkx>=2.8.0",
]

[project.optional-dependencies]
viz = ["umap-learn>=0.5.0"]
kaggle = ["kagglehub>=0.2.0"]
dev = ["pytest>=7.0.0", "black>=22.0.0", "isort>=5.10.0"]
full = ["umap-learn>=0.5.0", "kagglehub>=0.2.0"]

[project.urls]
"Homepage" = "https://github.com/example/integral-superposition"
"Bug Reports" = "https://github.com/example/integral-superposition/issues"
"Source" = "https://github.com/example/integral-superposition"

[tool.setuptools.packages.find]
where = ["."]
include = ["integral_superposition*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88
